pyproject.toml
[tool.poetry]
name = "holocron"
version = "0.4.0"
description = "An extendable static site generator powered by the Force. =/"
authors = ["Ihor Kalnytskyi <ihor@kalnytskyi.com>"]
license = "BSD-3-Clause"
readme = "README.rst"
homepage = "https://github.com/ikalnytskyi/holocron"
repository = "https://github.com/ikalnytskyi/holocron"
documentation = "https://holocron.readthedocs.io/"
keywords = ["static", "site", "blog", "generator", "markdown"]
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Terminals",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.5",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
]

[tool.poetry.dependencies]
python = "^3.5"
Jinja2 = "^2.10"
PyYAML = "^5.1"
Pygments = "^2.3"
python-dateutil = "^2.7"
jsonpointer = "^2.0"
Markdown = "^3.0"
docutils = "^0.14.0"
feedgen = "^0.7.0"
termcolor = "^1.1"
colorama = "^0.4.1"
mistletoe = "^0.7.1"
jsonschema = { version = "^3.0", extras = ["format"] }

[tool.poetry.dev-dependencies]
mock = "^2.0"
coverage = "^4.5"
pytest = "^4.1"
pytest-randomly = "^1.2"
flake8 = "^3.6"
xmltodict = "^0.11.0"
beautifulsoup4 = "^4.7"
untangle = "^1.1"
flake8-quotes = "^1.0"

[tool.poetry.scripts]
holocron = "holocron.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
